+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file. The version
 ### Fixed
 * When running a route in floor mode, it now adds single quotation marks around the route name so it supports spaces in the name.
 
+### Removed
+* Command `route-set-all-margins` and `route-set-all-max-speeds` (use `route-set-pos-option`)
+
 ## 1.4.0 - 2023-10-08
 
 ### Added